Move ClientConfig and ClientPlugin to client-api
Needed to move validateProtocolAndTransport from ClientPipeline to ClientConfig.
Needed to move withRequestOverride from ClientConfig to Client in order to keep
RequestOverrideConfig in client-core. Other option would have been to move
RequestOverrideConfig to client-api too, even though it is not an "api" that's
an extension point.
